Packaging task-instance.
