+++ src/prelude.rs
@@ -10,12 +10,13 @@
 //!
 //! The prelude may grow over time as additional items see ubiquitous use.
 
-pub use crate::entity::{get_timestamp, get_timestamp_micros, DelayTimer, DelayTimerBuilder};
 pub use crate::macros::*;
-pub use crate::timer::runtime_trace::task_handle::DelayTaskHandler;
 pub use crate::timer::task::TaskContext;
-pub use crate::timer::task::{Frequency, ScheduleIteratorTimeZone, Task, TaskBuilder};
 pub use crate::timer::timer_core::TimerEvent;
+pub use crate::timer::runtime_trace::task_instance::TaskInstancesChain;
+pub use crate::timer::runtime_trace::task_handle::DelayTaskHandler;
+pub use crate::entity::{get_timestamp, get_timestamp_micros, DelayTimer, DelayTimerBuilder};
+pub use crate::timer::task::{Frequency, ScheduleIteratorTimeZone, Task, TaskBuilder};
 pub use crate::utils::convenience::cron_expression_grammatical_candy::{
     CandyCron, CandyCronStr, CandyFrequency,
 };
@@ -30,6 +31,8 @@ pub use smol::spawn as async_spawn;
 pub use smol::unblock as unblock_spawn;
 
 pub(crate) use crate::entity::RuntimeKind;
+pub(crate) use crate::timer::runtime_trace::task_instance::TaskInstancesChainMaintainer;
+
 pub(crate) use crate::utils::parse::shell_command::{ChildGuard, ChildGuardList};
 pub(crate) use smol::channel::{Receiver as AsyncReceiver, Sender as AsyncSender};
 pub(crate) use smol::future::yield_now;
